requirements.txt
# ── Core ───────────────────────────────────────────
pydantic>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
faker>=19.0.0

# ── OpenEnv / Server ──────────────────────────────
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0

# ── Inference ─────────────────────────────────────
openai>=1.0.0
python-dotenv>=1.0.0

# ── Config / CLI ──────────────────────────────────
pyyaml>=6.0
rich>=13.0.0
typer>=0.9.0

# ── Testing ───────────────────────────────────────
pytest>=7.0.0
pytest-cov>=4.0.0
